Allow marshalling map[any]any for chart values
Without this change, if the rancher-config config map has nested objects
as chart values, e.g.:

data:
  rancher-webhook: |-
    auth:
      clientCA: foo

The resulting marshalled YAML will be a map[any]any, and the
unmarshalled object cannot be decoded by the catalog installer. This
change uses an alternative library so that nested values can be encoded.